projectweb/requirements.txt
Flask==2.0.1
Pillow==8.2.0
opencv-python==4.5.3.56
numpy==1.21.0
matplotlib==3.4.2
